requirements.txt
# DeepFake Defender Dependencies

# Web Framework
flask
werkzeug
waitress

# AI/ML
torch
torchvision
transformers
pillow
numpy
scipy
PyWavelets
joblib
scikit-learn
scikit-image
mediapipe

# Video detection
opencv-python
onnx
onnx2pytorch

# C2PA metadata reader
c2pa-python
